requirements.txt
cassandra-driver==3.22.0
numpy==1.18.3
pandas==1.0.3
geopy==1.21.0